requirements.txt
Flask==2.3.2
Werkzeug==2.3.6
Jinja2==3.1.2
python-gnupg==0.5.1
click==8.1.3
itsdangerous==2.1.2
MarkupSafe==2.1.3 